UltraVNC: SVN revision 1068

+++ ica/win32/rdr/ZlibOutStream.cxx
@@ -19,9 +19,9 @@
 #include "ZlibOutStream.h"
 #include "Exception.h"
 #ifdef IPP
-#include "../ipp_zlib/src/zlib/zlib.h"
+#include "../ipp_zlib/src/zlib-1.2.5/zlib.h"
 #else
-#include "../zlib/zlib.h"
+#include "../zlib-1.2.5/zlib.h"
 #endif
 
 using namespace rdr;
@@ -62,7 +62,7 @@ void ZlibOutStream::setUnderlying(OutStream* os)
 
 int ZlibOutStream::length()
 {
-  return offset + ptr - start;
+  return (int)(offset + ptr - start);
 }
 
 void ZlibOutStream::flush()
